requirements.txt
agate==1.9.1
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.4.0
babel==2.17.0
backports.tarfile==1.2.0
blinker==1.9.0
boto3==1.41.5
botocore==1.41.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.0
daff==1.4.2
dbt-adapters==1.20.1
dbt-common==1.36.0
dbt-core==1.11.0rc1
dbt-extractor==0.6.0
dbt-protos==1.0.397
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.10.3
debugpy==1.8.17
decorator==5.2.1
deepdiff==8.6.1
executing==2.2.1
Faker==40.1.2
filelock==3.20.0
findspark==2.0.1
Flask==3.1.3
gunicorn==25.1.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keyring==25.7.0
leather==0.4.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib-inline==0.2.1
more-itertools==10.8.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.6
numpy==2.4.2
orderly-set==5.5.0
packaging==25.0
pandas==3.0.1
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.31.1
psutil==7.1.3
pure_eval==0.2.3
py4j==0.10.9.7
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
s3transfer==0.15.0
six==1.17.0
snowflake-connector-python==3.18.0
snowflake-snowpark-python==1.42.0
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
tomlkit==0.13.3
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.6
zipp==3.23.0
